requirements.txt
faiss-cpu
numpy
openai
pdfplumber
matplotlib
seaborn
scikit-learn
networkx
